# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DownloadError {
  error: string;
  detail?: string;
}

export type DownloadHlsAsMp4Params = {
/**
 * Public HTTP(S) URL of the HLS playlist.
 */
url: string;
/**
 * Fast remux when the source codecs are MP4-compatible, or a broader compatibility conversion.
 */
mode?: DownloadHlsAsMp4Mode;
};

export type DownloadHlsAsMp4Mode = typeof DownloadHlsAsMp4Mode[keyof typeof DownloadHlsAsMp4Mode];


export const DownloadHlsAsMp4Mode = {
  compatible: 'compatible',
  fast: 'fast',
} as const;

export type ProxyHlsResourceParams = {
/**
 * Public HTTP(S) URL of an HLS playlist, segment, or key.
 */
url: string;
};